added mac os support for ansible and iOS devices

+++ zebrunner.sh
@@ -3,8 +3,47 @@
 ## shellcheck disable=SC1091
 #source patch/utility.sh
 
+  replace() {
+    #TODO: https://github.com/zebrunner/zebrunner/issues/328 organize debug logging for setup/replace
+    file=$1
+    #echo "file: $file"
+    content=$(<$file) # read the file's content into
+    #echo "content: $content"
+
+    old=$2
+    #echo "old: $old"
+
+    new=$3
+    #echo "new: $new"
+    content=${content//"$old"/$new}
+
+    #echo "content: $content"
+
+    printf '%s' "$content" >$file    # write new content to disk
+  }
+
   setup() {
-    cp roles/devices/vars/main.yml.original roles/devices/vars/main.yml
+    if [ -f roles/devices/vars/main.yml ]; then
+      echo "Skip roles/devices/vars/main.yml override as file already present!"
+    else
+      cp roles/devices/vars/main.yml.original roles/devices/vars/main.yml
+    fi
+
+    if [ -f roles/mac-devices/vars/main.yml ]; then
+      echo "Skip roles/mac-devices/vars/main.yml override as file already present!"
+    else
+      cp roles/mac-devices/vars/main.yml.original roles/mac-devices/vars/main.yml
+    fi
+
+    if [ -d $HOME/Library/LaunchAgents ] && [ ! -f $HOME/Library/LaunchAgents/ZebrunnerDevicesListener.plist ]; then
+      # register devices manager to manage attach/reboot actions
+      cp roles/mac-devices/templates/ZebrunnerDevicesListener.plist $HOME/Library/LaunchAgents/ZebrunnerDevicesListener.plist
+      replace $HOME/Library/LaunchAgents/ZebrunnerDevicesListener.plist "working_dir_value" "${BASEDIR}"
+      replace $HOME/Library/LaunchAgents/ZebrunnerDevicesListener.plist "user_value" "$USER"
+
+      launchctl load $HOME/Library/LaunchAgents/ZebrunnerDevicesListener.plist
+    fi
+
     #TODO: switch to master branch after oficial release and merge
     echo "Follow https://github.com/zebrunner/mcloud-agent/tree/develop#run-ansible-playbook to deploy MCloud agent services!"
   }
@@ -22,6 +61,11 @@
       exit
     fi
 
+    if [ -f $HOME/Library/LaunchAgents/ZebrunnerDevicesManager.plist ]; then
+      launchctl unload $HOME/Library/LaunchAgents/ZebrunnerDevicesManager.plist
+      rm -f $HOME/Library/LaunchAgents/ZebrunnerDevicesManager.plist
+    fi
+
     down
 
     sudo rm -f /usr/local/bin/zebrunner-farm
@@ -29,6 +73,7 @@
     sudo rm -f /etc/udev/rules.d/90_mcloud.rules
     # restore original main.yml
     rm -f roles/devices/vars/main.yml
+    rm -f roles/mac-devices/vars/main.yml
 
     docker volume rm appium-storage-volume
     docker volume rm mcloud-storage-volume
@@ -73,8 +118,9 @@
     cp /usr/local/bin/mcloud-devices.txt backup/
     cp /etc/udev/rules.d/90_mcloud.rules backup/
     cp roles/devices/vars/main.yml roles/devices/vars/main.yml.bak
+    cp roles/mac-devices/vars/main.yml roles/mac-devices/vars/main.yml.bak
 
-    if [ -f backup/zebrunner-farm ] && [ -f backup/mcloud-devices.txt ] && [ -f backup/90_mcloud.rules ] && [ -f roles/devices/vars/main.yml.bak ]; then
+    if [ -f backup/zebrunner-farm ] && [ -f backup/mcloud-devices.txt ] && [ -f backup/90_mcloud.rules ] && [ -f roles/devices/vars/main.yml.bak ] && [ -f roles/mac-devices/vars/main.yml.bak ]; then
       echo "MCloud backup succeed."
     else
       echo_warning "MCloud backup failed!"
@@ -92,11 +138,12 @@
     sudo cp backup/mcloud-devices.txt /usr/local/bin/mcloud-devices.txt
     sudo cp backup/90_mcloud.rules /etc/udev/rules.d/90_mcloud.rules
     cp roles/devices/vars/main.yml.bak roles/devices/vars/main.yml
+    cp roles/mac-devices/vars/main.yml.bak roles/mac-devices/vars/main.yml
 
     # reload udevadm rules
     sudo udevadm control --reload-rules
 
-    if [ -f /usr/local/bin/zebrunner-farm ] && [ -f /usr/local/bin/mcloud-devices.txt ] && [ -f /etc/udev/rules.d/90_mcloud.rules ] && [ -f roles/devices/vars/main.yml ]; then
+    if [ -f /usr/local/bin/zebrunner-farm ] && [ -f /usr/local/bin/mcloud-devices.txt ] && [ -f /etc/udev/rules.d/90_mcloud.rules ] && [ -f roles/devices/vars/main.yml ] && [ -f roles/mac-devices/vars/main.yml ]; then
       echo "MCloud restore succeed."
     else
       echo_warning "MCloud restore failed!"
